Moving ActionArea by dragging polygon
This is harder than i thought due to something that looks like a bug?
When i set Canvas.setleft(polygon, ##); it moves the polygon leftwrad ##
amount, but the Points of the polygon arne't updated for some reason...
so I had to do a hack work around. It seems to work quite smoothly so
far...